+++ p2panda-auth/src/lib.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+mod group;
+pub mod group_crdt;
+mod traits;