+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Highlights are marked with a pancake 🥞
 ### Added
 
 - Expose sqlite migrations [#744](https://github.com/p2panda/p2panda/pull/744)
+- Introduce core group CRDT for auth [#748](https://github.com/p2panda/p2panda/pull/748)
 
 ## [0.3.1] - 14/04/2025